+++ docs/src/pages/docs.astro
@@ -1,8 +1,9 @@
 ---
-import Layout from "@/layouts/layout.astro";
-import { Navbar } from "@/components/navbar";
 import FunctionCard from "@/components/function-card.astro";
-import { FUNCTIONS } from "@/data";
+import { Navbar } from "@/components/navbar";
+import Layout from "@/layouts/layout.astro";
+import { flatten, groupBy, map, pipe, values } from "remeda";
+import DATA from "../../build/data.json";
 ---
 
 <Layout>
@@ -17,7 +18,15 @@ import { FUNCTIONS } from "@/data";
 
     <main class="h-full py-6">
       <div class="space-y-6">
-        {FUNCTIONS.map((func) => <FunctionCard func={func} />)}
+        {
+          pipe(
+            DATA,
+            groupBy((func) => func.category),
+            values,
+            flatten(),
+            map((func) => <FunctionCard func={func} />),
+          )
+        }
       </div>
     </main>
   </div>
